+++ doc/changes.md
@@ -7,6 +7,7 @@
 * NEW: pre-flight checks for output format
 * NEW: add `type:TYPE` option to the dnsstat output formatter, automatically add this for new DNS measurements
 * CHANGED: update to go1.21 and update dependencies
+* NEW: add a progress indicator to the dnsstat output formatter
 
 ## v0.5.0